backend/requirements.txt
# --- Core API Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
slowapi>=0.1.9

# --- Database & Cache ---
psycopg2-binary>=2.9.9
pgvector>=0.2.4
redis>=5.0.0
sqlalchemy>=2.0.23

# --- AI & Data Processing ---
sentence-transformers>=2.2.2
cohere>=4.38
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
pypdf2>=3.0.1
numpy>=1.24.0
torch>=2.1.0
google-generativeai>=0.3.0
celery>=5.3.4
alembic>=1.13.0
pyspellchecker>=0.7.2
python-Levenshtein>=0.23.0

# --- File Format Parsers ---
python-docx>=1.1.0
python-pptx>=0.6.23

# --- Authentication ---
python-jose>=3.3.0
# Pinned to fix version conflict causing registration errors
passlib==1.7.4
bcrypt==3.2.0
cryptography>=41.0.0
duckduckgo-search>=4.1.0
